Make the refactoring build properly.
- A *lot* of left behind garbage that needs cleaned up, but getting there.
