+++ src/komodo/komodo_rpcblockchain.h
@@ -16,11 +16,54 @@
 #ifndef komodo_rpcblockchain_h
 #define komodo_rpcblockchain_h
 
+namespace komodo {
 
-int32_t komodo_MoMdata(int32_t *notarized_htp,uint256 *MoMp,uint256 *kmdtxidp,int32_t height,uint256 *MoMoMp,int32_t *MoMoMoffsetp,int32_t *MoMoMdepthp, int32_t *kmdstartip, int32_t *kmdendip);
+struct notarized_checkpoint
+{
+    uint256 notarized_hash,notarized_desttxid,MoM,MoMoM;
+    int32_t nHeight,notarized_height,MoMdepth,MoMoMdepth,MoMoMoffset,kmdstarti,kmdendi;
+} *NPOINTS;
+
+int32_t NUM_NPOINTS,last_NPOINTSi,NOTARIZED_HEIGHT,NOTARIZED_MOMDEPTH,KOMODO_NEEDPUBKEYS;
 
 
-int32_t komodo_MoM(int32_t *notarized_heightp,uint256 *MoMp,uint256 *kmdtxidp,int32_t nHeight,uint256 *MoMoMp,int32_t *MoMoMoffsetp,int32_t *MoMoMdepthp,int32_t *kmdstartip,int32_t *kmdendip)
+struct notarized_checkpoint *komodo_npptr(uint64_t height)
+{
+    int i; struct notarized_checkpoint *np = 0;
+    for (i=NUM_NPOINTS-1; i>=0; i--)
+    {
+        np = &NPOINTS[i];
+        if ( np->MoMdepth > 0 && height > np->notarized_height-np->MoMdepth && height <= np->notarized_height )
+            return(np);
+    }
+    return(0);
+}
+
+
+int32_t komodo_MoMdata(int32_t *notarized_htp,uint256 *MoMp,uint256 *kmdtxidp,int32_t height,uint256 *MoMoMp, int32_t *MoMoMoffsetp, int32_t *MoMoMdepthp, int32_t *kmdstartip, int32_t *kmdendip)
+{
+    struct notarized_checkpoint *np = 0;
+    if ( (np= komodo_npptr(height)) != 0 )
+    {
+        *notarized_htp = np->notarized_height;
+        *MoMp = np->MoM;
+        *kmdtxidp = np->notarized_desttxid;
+        *MoMoMp = np->MoMoM;
+        *MoMoMoffsetp = np->MoMoMoffset;
+        *MoMoMdepthp = np->MoMoMdepth;
+        *kmdstartip = np->kmdstarti;
+        *kmdendip = np->kmdendi;
+        return(np->MoMdepth);
+    }
+    *notarized_htp = *MoMoMoffsetp = *MoMoMdepthp = *kmdstartip = *kmdendip = 0;
+    memset(MoMp,0,sizeof(*MoMp));
+    memset(MoMoMp,0,sizeof(*MoMoMp));
+    memset(kmdtxidp,0,sizeof(*kmdtxidp));
+    return(0);
+}
+
+
+static int32_t komodo_MoM(int32_t *notarized_heightp,uint256 *MoMp,uint256 *kmdtxidp,int32_t nHeight,uint256 *MoMoMp,int32_t *MoMoMoffsetp,int32_t *MoMoMdepthp,int32_t *kmdstartip,int32_t *kmdendip)
 {
     int32_t depth;
     int32_t notarized_ht;
@@ -46,6 +89,21 @@ int32_t komodo_MoM(int32_t *notarized_heightp,uint256 *MoMp,uint256 *kmdtxidp,in
     }
     return(depth);
 }
+
+int32_t komodo_prevMoMheight()
+{
+    static uint256 zero;
+    int32_t i; struct notarized_checkpoint *np = 0;
+    for (i=NUM_NPOINTS-1; i>=0; i--)
+    {
+        np = &NPOINTS[i];
+        if ( np->MoM != zero )
+            return(np->notarized_height);
+    }
+    return(0);
+}
+
+} // namespace komodo
 
 #endif /* komodo_rpcblockchain_h */
